+++ src/worker/WorkThread.h
@@ -12,6 +12,20 @@
 #include "FileContext.h"
 #include "ItemContext.h"
 
+bool ProgresssLoop(CFileContext* pContext, CPipe &Stderr, int &nProgress);
+
+bool ReadLoop(CPipeContext* pContext);
+
+bool WriteLoop(CPipeContext* pContext);
+
+bool ConvertFileUsingConsole(CFileContext* pContext);
+
+bool ConvertFileUsingPipes(CFileContext* pContext);
+
+bool ConvertFileUsingOnlyPipes(CFileContext* pDecoderContext, CFileContext* pEncoderContext);
+
 bool ConvertItem(CItemContext* pContext);
 
-int WorkThread(void *param);
+bool ConvertLoop(CWorkerContext* pWorkerContext);
+
+void Convert(CWorkerContext* pWorkerContext);